refactor(dao): remove legacy dao package
- delete unused classes
