+++ src/main/java/chess/dao/EventDao.java
@@ -1,40 +1,44 @@
 package chess.dao;
 
 import chess.domain.event.Event;
-import java.util.ArrayList;
 import java.util.List;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public class EventDao {
 
     private static final String TABLE_NAME = "event";
 
-    public List<Event> findAllByGameId(int gameId) {
-        final String sql = addTable("SELECT type, description FROM %s WHERE game_id = ?");
-        final ResultReader reader = new StatementExecutor(sql).setInt(gameId)
-                .executeQuery();
-        try (reader) {
-            return readAllEvents(reader);
-        }
+    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
+    public EventDao(NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+        this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
     }
 
-    private List<Event> readAllEvents(ResultReader reader) {
-        List<Event> pieces = new ArrayList<>();
-        while (reader.hasNextRow()) {
-            String eventType = reader.readStringAt("type");
-            String description = reader.readStringAt("description");
-            Event event = Event.of(eventType, description);
-            pieces.add(event);
-        }
-        return pieces;
+    private final RowMapper<Event> eventRowMapper = (resultSet, rowNum) ->
+            Event.of(resultSet.getString("type"),
+                    resultSet.getString("description"));
+
+    public List<Event> findAllByGameId(int gameId) {
+        final String sql = addTable("SELECT type, description FROM %s WHERE game_id = :game_id");
+        SqlParameterSource paramSource = new MapSqlParameterSource("game_id", gameId);
+        return namedParameterJdbcTemplate.query(sql, paramSource, eventRowMapper);
     }
 
     public void save(int gameId, Event event) {
-        final String sql = addTable("INSERT INTO %s (game_id, type, description) VALUES (?, ?, ?)");
+        final String sql = addTable("INSERT INTO %s (game_id, type, description) "
+                + "VALUES (:game_id, :type, :description)");
 
-        new StatementExecutor(sql).setInt(gameId)
-                .setString(event.getType())
-                .setString(event.getDescription())
-                .executeCommand();
+        MapSqlParameterSource paramSource = new MapSqlParameterSource();
+        paramSource.addValue("game_id", gameId);
+        paramSource.addValue("type", event.getType());
+        paramSource.addValue("description", event.getDescription());
+
+        namedParameterJdbcTemplate.update(sql, paramSource);
     }
 
     protected String addTable(String sql) {
