+++ src/test/java/chess/dao/EventDaoTest.java
@@ -5,35 +5,41 @@ import static org.assertj.core.api.Assertions.assertThat;
 import chess.domain.event.Event;
 import chess.domain.event.MoveEvent;
 import java.util.List;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 
 @SuppressWarnings("NonAsciiCharacters")
+@JdbcTest
 class EventDaoTest {
 
     private static final String TEST_TABLE = "event_test";
-    private static final String SETUP_TEST_DB_SQL = String.format("INSERT INTO %s (game_id, type, description) "
-            + "VALUES (1, 'MOVE', 'a2 a4'), (1, 'MOVE', 'a7 a5'), (2, 'MOVE', 'a2 a3')", TEST_TABLE);
 
-    private static final String CLEANSE_TEST_DB_SQL = String.format("TRUNCATE TABLE %s", TEST_TABLE);
+    @Autowired
+    private JdbcTemplate jdbcTemplate;
 
-    private final EventDao dao = new EventDao() {
-        @Override
-        protected String addTable(String sql) {
-            return String.format(sql, TEST_TABLE);
-        }
-    };
+    @Autowired
+    private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
+
+    private EventDao dao;
 
     @BeforeEach
     void setUp() {
-        cleanUp();
-        new StatementExecutor(SETUP_TEST_DB_SQL).executeCommand();
-    }
+        dao = new EventDao(namedParameterJdbcTemplate) {
+            @Override
+            protected String addTable(String sql) {
+                return String.format(sql, TEST_TABLE);
+            }
+        };
+        jdbcTemplate.execute("DROP TABLE event_test IF EXISTS");
+        jdbcTemplate.execute("CREATE TABLE event_test(game_id  BIGINT NOT NULL, "
+                + "type VARCHAR(20) NOT NULL, description VARCHAR(20))");
 
-    @AfterEach
-    void cleanUp() {
-        new StatementExecutor(CLEANSE_TEST_DB_SQL).executeCommand();
+        jdbcTemplate.execute("INSERT INTO event_test (game_id, type, description) "
+                + "VALUES (1, 'MOVE', 'a2 a4'), (1, 'MOVE', 'a7 a5'), (2, 'MOVE', 'a2 a3')");
     }
 
     @Test
